separate core and router power gating

+++ src/flovtrafficmanager.hpp
@@ -50,9 +50,9 @@ private:
   /* ==== Power Gate - End ==== */
 
   /* ==== Power Gate - Begin ==== */
-  vector<vector<uint64_t> > _idle_cycles;
-  vector<uint64_t> _router_idle_periods;
-  vector<vector<uint64_t> > _overall_idle_cycles;
+  vector<vector<int> > _idle_cycles;
+  vector<int> _router_idle_periods;
+  vector<vector<int> > _overall_idle_cycles;
   vector<bool> _wakeup_handshake_latency;
   /* ==== Power Gate - End ==== */
   // ============ Internal methods ============
